+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ## [Unreleased 2.x]
 ### Added
+- Added support for indexing and search index settings ([#667](https://github.com/opensearch-project/opensearch-java/pull/667))
 
 ### Dependencies